+++ src/Ext/BuildingType/Body.h
@@ -129,6 +129,11 @@ public:
 
 	Nullable<int> DeployFireDelay;
 
+	Valueable<int> ActiveAnim_MoneyAmount;
+	Valueable<int> ActiveAnimTwo_MoneyAmount;
+	Valueable<int> ActiveAnimThree_MoneyAmount;
+	Valueable<int> ActiveAnimFour_MoneyAmount;
+
 	// Ares 0.2
 	Valueable<bool> CloningFacility;
 
@@ -225,6 +230,10 @@ public:
 		, SetTabBySelecting { -1 }
 		, RevealToAll_Radius {}
 		, DeployFireDelay {}
+		, ActiveAnim_MoneyAmount { -1 }
+		, ActiveAnimTwo_MoneyAmount { -1 }
+		, ActiveAnimThree_MoneyAmount { -1 }
+		, ActiveAnimFour_MoneyAmount { -1 }
 
 		// Ares 0.2
 		, CloningFacility { false }