+++ docs/Whats-New.md
@@ -702,6 +702,7 @@ HideShakeEffects=false           ; boolean
 - [Reload speed adjustment on promotion](New-or-Enhanced-Logics.md#reload-speed-adjustment-on-promotion) (by Nuke)
 - Allowed `(Pre)ProductionAnim` animations to use `Powered` & `PoweredLight/Effect/Special` keys (by Noble_Fish)
 - Customize ivan bomb visibility (by NetsuNegi)
+- Money switch building active animation (by dh381)
 
 #### Vanilla fixes:
 - Fixed sidebar not updating queued unit numbers when adding or removing units when the production is on hold (by CrimRecya)